requirements.txt
fastapi
uvicorn[standard]
docker
GitPython
python-dotenv
huggingface_hub
pydantic